A bit of refactoring now that we're entirely on JSON ast
