+++ src/test/errors-analysis.test.ts
@@ -795,7 +795,7 @@ test(
     assert.match(
       done.stderr,
       `
-❌ [../bar:b] Invalid JSON in package.json file in ${pathlib.resolve(
+❌ [../bar:b] Invalid JSON syntax in package.json file in ${pathlib.resolve(
         rig.temp,
         'bar'
       )}
