+++ src/test/parallelism.test.ts
@@ -8,9 +8,8 @@ import {suite} from 'uvu';
 import * as assert from 'uvu/assert';
 import {timeout, wait} from './util/uvu-timeout.js';
 import {WireitTestRig} from './util/test-rig.js';
-import {PackageJson} from '../util/package-json-reader.js';
+import { PackageJson } from "./util/package-json";
 import * as os from 'os';
-import {addAst} from './util/add-ast.js';
 
 const test = suite<{rig: WireitTestRig}>();
 
@@ -143,14 +142,14 @@ test(
     // that default.
     const n = os.cpus().length * 10;
     const depNames: string[] = [];
-    const packageJson: PackageJson = addAst({
+    const packageJson: PackageJson = {
       scripts: {
         main: 'wireit',
       },
       wireit: {
         main: {command: main.command, dependencies: depNames},
       },
-    });
+    };
     const commands = [];
     const invocations = [];
     for (let i = 0; i < n; i++) {
