+++ src/util/package-json-reader.ts
@@ -12,34 +12,14 @@ import { PlaceholderConfig } from '../analyzer.js';
 export const astKey = Symbol('ast');
 
 /**
- * A raw package.json JSON object, including the special "wireit" section.
- */
-export interface PackageJson {
-  name?: string;
-  version?: string;
-  scripts?: {[scriptName: string]: string};
-  wireit?: {
-    [scriptName: string]: {
-      command?: string;
-      dependencies?: string[];
-      files?: string[];
-      output?: string[];
-      clean?: boolean | 'if-file-deleted';
-      packageLocks?: string[];
-    };
-  };
-  [astKey]: AstNode;
-}
-
-/**
  * Reads package.json files and caches them.
  */
 export class CachingPackageJsonReader {
-  readonly #cache = new Map<string, PackageJson>();
+  readonly #cache = new Map<string, AstNode>();
 
-  async read(packageDir: string, placeholder: PlaceholderConfig): Promise<PackageJson> {
-    let packageJson = this.#cache.get(packageDir);
-    if (packageJson === undefined) {
+  async read(packageDir: string, placeholder: PlaceholderConfig): Promise<AstNode> {
+    let ast = this.#cache.get(packageDir);
+    if (ast === undefined) {
       const packageJsonPath = pathlib.resolve(packageDir, 'package.json');
       let packageJsonStr: string;
       try {
@@ -50,19 +30,10 @@ export class CachingPackageJsonReader {
         }
         throw error;
       }
-      try {
-        packageJson = JSON.parse(packageJsonStr) as PackageJson;
-      } catch (error) {
-        throw new CachingPackageJsonReaderError('invalid-package-json');
-      }
-      const ast = parseTree(packageJsonStr, placeholder);
-      if (ast === undefined) {
-        throw new CachingPackageJsonReaderError('invalid-package-json');
-      }
-      packageJson[astKey] = ast as AstNode;
-      this.#cache.set(packageDir, packageJson);
+      ast = parseTree(packageJsonStr, placeholder);
+      this.#cache.set(packageDir, ast);
     }
-    return packageJson;
+    return ast;
   }
 }
